SwapPairContract wrapper added + some contracts files moved to other directories
